% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deFinetti.plot}
\alias{deFinetti.plot}
\title{Plot a de Finetti diagram}
\usage{
deFinetti.plot(x, file = "", main = "", vertexlab = colnames(x),
  cex = 1, cex.main = 2, cex.lab = 2, pch = 19, markerlab = NULL,
  markerpos = 1, markercex = 1, markercol = "black", markerbg = "black",
  hwcurve = TRUE, with_F_color = FALSE, without_markers = FALSE,
  pdf_out = FALSE, ...)
}
\arguments{
\item{x}{a numeric \code{matrix}}

\item{file}{a string of output filename}

\item{main}{a string of axis label}

\item{vertexlab}{a string \code{vector} of vertex labels}

\item{cex}{a numeric of character expansion factor}

\item{cex.main}{a numeric of magnification for titles relative to \code{cex}}

\item{cex.lab}{a numeric of magnification for labels relative to \code{cex}}

\item{pch}{a symbol for plotting markers}

\item{markerlab}{a string \code{vector} of marker labels}

\item{markerpos}{a position specifier for marker labels}

\item{markercex}{a numeric of magnification for markers relative to \code{cex}}

\item{markercol}{a color for plotting marker symbols}

\item{markerbg}{a background (fill) color for plotting marker symbols given by \code{pch = 21:25}}

\item{hwcurve}{a logical indicating whether to draw a curve of Hardy-Weinberg Equilibrium}

\item{with_F_color}{a logical indicating whether to draw colors based on F-statistics}

\item{without_markers}{a logical indicating whether to plot markers}

\item{pdf_out}{a logical indicating whether to output a diagram}

\item{...}{arguments to be passed to \code{plot}}
}
\description{
Plot a de Finetti diagram
}

